switch Feed.js to typescript
